hw6

+++ controllers/contacts-controller.js
@@ -3,13 +3,19 @@ import {HttpError} from "../helpers/index.js"
 import controllerWrapper from "../decorators/controllerWrapper.js";
 
 export const getAll = async (req, res) => {
-    const result = await Contact.find();
+    const {_id: owner} = req.user;
+    const query = { owner } 
+    const result = await Contact.find(query).populate("owner", "email subscribtion");
+    if(result.length === 0){
+        throw HttpError(404, `No contacts added`);            
+    }
     res.json(result);
 };
 
 export const getById = async (req, res) => {
-    const {id} = req.params;
-    const result = await Contact.findById(id)
+    const {_id: owner} = req.user
+	const { id } = req.params;
+	const result = await Contact.findOne({_id: id, owner});
     if (!result) {
         throw HttpError(404, `Contact with id=${id} not found`)
     }
@@ -21,13 +27,15 @@ export const addContact = async (req, res) => {
         if (error) {
             throw HttpError(400, error.message)
         }
-        const result = await Contact.create(req.body)
+        const {_id: owner} = req.user
+        const result = await Contact.create({...req.body, owner})
         res.status(201).json(result)
 }
 
 export const updateById = async (req, res) => {
+        const {_id: owner} = req.user
         const {id} = req.params;
-        const result = await Contact.findByIdAndUpdate(id, req.body)
+        const result = await Contact.findByIdAndUpdate({_id: id, owner}, req.body)
         if (!result) {
             throw HttpError(404, `Contact with id=${id} not found`)
         }
@@ -36,7 +44,8 @@ export const updateById = async (req, res) => {
 
 export const deleteById = async (req, res) => {
         const {id} = req.params;
-        const result = await Contact.findByIdAndDelete(id)
+        const {_id: owner} = req.user
+        const result = await Contact.findByIdAndDelete({_id: id, owner})
         if (!result) {
             throw HttpError(404, `Contact with id=${id} not found`)
         }
@@ -52,4 +61,3 @@ export default {
 	deleteContact: controllerWrapper(deleteById),
 	updateContactById: controllerWrapper(updateById),
 };
-
